pyproject.toml
[project]
name = "mlops-e1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.13",
    "asyncio>=3.4.3",
    "fastapi[standard]>=0.115.7",
    "polars>=1.23.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
